src/test/javascript/spec/app/entities/sirket-bilgileri/sirket-bilgileri.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { SirketBilgileriService } from 'app/entities/sirket-bilgileri/sirket-bilgileri.service';
import { ISirketBilgileri, SirketBilgileri } from 'app/shared/model/sirket-bilgileri.model';

describe('Service Tests', () => {
  describe('SirketBilgileri Service', () => {
    let injector: TestBed;
    let service: SirketBilgileriService;
    let httpMock: HttpTestingController;
    let elemDefault: ISirketBilgileri;
    let expectedResult: ISirketBilgileri | ISirketBilgileri[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(SirketBilgileriService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new SirketBilgileri(0, 'AAAAAAA', 'AAAAAAA');
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a SirketBilgileri', () => {
        const returnedFromService = Object.assign(
          {
            id: 0,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new SirketBilgileri()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a SirketBilgileri', () => {
        const returnedFromService = Object.assign(
          {
            sirketAdi: 'BBBBBB',
            sirketAdresi: 'BBBBBB',
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of SirketBilgileri', () => {
        const returnedFromService = Object.assign(
          {
            sirketAdi: 'BBBBBB',
            sirketAdresi: 'BBBBBB',
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a SirketBilgileri', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
